requirements.txt
Scrapy==1.5.0
python-djvulibre==0.8.1
python-slugify==1.2.5
ScrapyElasticSearch==0.9.1
furl==1.1
Cython==0.28.3
dateparser==0.7.0
pdfminer.six==20170720
